+++ packages/connectivity/src/scp-cf/destination/destination-from-env.ts
@@ -162,23 +162,3 @@ function destinationHasName(
 ): destination is DestinationWithName {
   return !!destination.name;
 }
-
-/**
- * Set a given destination in the `destinations` environment variable.
- *
- * Throws an error if a destination with the same name as the given test destination already exists.
- * @param destination - A destination to add to the `destinations` environment variable
- */
-export function registerDestination(destination: DestinationWithName): void {
-  const currentDestinations = getDestinationsFromEnv();
-  const existingNames = new Set<string>(
-    currentDestinations.filter(destinationHasName).map(dest => dest.name)
-  );
-  if (!destination.name || !destination.url) {
-    throw Error(
-      'The registerDestination function requires a destination name and url.'
-    );
-  }
-  validateNameAvailable(destination.name, existingNames);
-  setDestinationsInEnv([...currentDestinations, destination]);
-}
